- optimized float-to-int conversion
git-svn-id: https://horde3d.svn.sourceforge.net/svnroot/horde3d/trunk@118 768fa460-af4a-0410-8817-c9c6fcdc2709

+++ Horde3D/Source/ColladaConverter/converter.cpp
@@ -603,7 +603,7 @@ void Converter::processMeshes( ColladaDocument &doc, bool optimize )
 						for( unsigned int l = 0; l < vertWeights.size(); ++l )
 						{
 							if( l == 4 ) break;
-							v.weights[l] = skin->weightArray->floatArray[(int)vertWeights[l].weight];
+							v.weights[l] = skin->weightArray->floatArray[vertWeights[l].weight];
 							v.joints[l] = jointLookup[vertWeights[l].joint];
 						}
